Move users state into module
