+++ src/immutable/state/index.ts
@@ -1,7 +1,6 @@
 import { Record } from 'immutable';
 import { RouterState } from 'connected-react-router';
 
-import { UsersStateRecord } from './users';
 import { ADMIN_NAMESPACE } from '../../modules/admin/constants';
 import { AdminStateRecord } from '../../modules/admin/state/index';
 import { CORE_NAMESPACE } from '../../modules/core/constants';
@@ -9,8 +8,7 @@ import { CoreStateRecord } from '../../modules/core/state/index';
 import { DASHBOARD_NAMESPACE } from '../../modules/dashboard/constants';
 import { DashboardStateRecord } from '../../modules/dashboard/state/index';
 import { USERS_NAMESPACE } from '../../modules/users/constants';
-
-export * from './users';
+import { UsersStateRecord } from '../../modules/users/state/index';
 
 export interface RootStateProps {
   admin: AdminStateRecord;
